monitoring: varius changes
* split monitoring into cli, api and views files.
* adds missing cli for Redis and ES.
