+++ rero_mef/monitoring/cli.py
@@ -1,0 +1,160 @@
+# -*- coding: utf-8 -*-
+#
+# RERO MEF
+# Copyright (C) 2022 RERO
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, version 3 of the License.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+"""Monitoring utilities."""
+
+import click
+from flask import current_app
+from flask.cli import with_appcontext
+from invenio_db import db
+from invenio_search import current_search_client
+from redis import Redis
+
+from .api import Monitoring
+from .utils import DB_CONNECTION_COUNTS_QUERY, DB_CONNECTION_QUERY
+
+
+@click.group()
+def monitoring():
+    """Monitoring commands."""
+
+
+@monitoring.command('es_db_counts')
+@click.option('-m', '--missing', 'missing', is_flag=True, default=False,
+              help='display missing pids')
+@with_appcontext
+def es_db_counts_cli(missing):
+    """Print ES and DB counts.
+
+    Prints a table representation of database and elasticsearch counts.
+    Columes:
+    1. database count minus elasticsearch count
+    2. document type
+    3. database count
+    4. elasticsearch index
+    5. elasticsearch count
+    """
+    missing_doc_types = []
+    mon = Monitoring()
+    msg_head = f'DB - ES  {"type":>6} {"count":>10}'
+    msg_head += f'  {"index":>25} {"count_es":>10}\n'
+    msg_head += f'{"":-^64s}'
+    click.echo(msg_head)
+    info = mon.info(with_deleted=False, difference_db_es=False)
+    for doc_type in sorted(info):
+        db_es = info[doc_type].get('db-es', '')
+        msg = f'{db_es:>7}  {doc_type:>6} {info[doc_type].get("db", ""):>10}'
+        index = info[doc_type].get('index', '')
+        if index:
+            msg += f'  {index:>25} {info[doc_type].get("es", ""):>10}'
+        if db_es not in [0, '']:
+            click.secho(msg, fg='red')
+        else:
+            click.echo(msg)
+        if missing and index:
+            missing_doc_types.append(doc_type)
+    for missing_doc_type in missing_doc_types:
+        mon.print_missing(missing_doc_type)
+
+
+@monitoring.command('mef_counts')
+@with_appcontext
+def mef_counts_cli():
+    """Print MEF counts.
+
+    Prints a table representation of MEF counts.
+    Columes:
+    1. MEF count minus database count
+    2. document type
+    3. database count
+    5. MEF count
+    """
+    mon = Monitoring()
+    msg_head = f'MEF - DB  {"type":>6} {"DB":>10}  {"MEF":>10}'
+    click.echo(msg_head)
+    for entity, data in mon.check_mef().items():
+        mef_db = data.get('mef-db', '')
+        db = data.get('db', '')
+        mef = data.get('mef', '')
+        msg = f'{mef_db:>8}  {entity:>6} {db:>10}  {mef:>10}'
+        if mef_db not in [0, '']:
+            click.secho(msg, fg='red')
+        else:
+            click.echo(msg)
+
+
+@monitoring.command('es_db_missing')
+@click.argument('doc_type')
+@with_appcontext
+def es_db_missing_cli(doc_type):
+    """Print missing pids informations."""
+    Monitoring().print_missing(doc_type)
+
+
+@monitoring.command()
+@with_appcontext
+def es():
+    """Displays elastic search cluster info."""
+    for key, value in current_search_client.cluster.health().items():
+        click.echo(f'{key:<33}: {value}')
+
+
+@monitoring.command()
+@with_appcontext
+def redis():
+    """Displays redis info."""
+    url = current_app.config.get('ACCOUNTS_SESSION_REDIS_URL',
+                                 'redis://localhost:6379')
+    redis = Redis.from_url(url)
+    for key, value in redis.info().items():
+        click.echo(f'{key:<33}: {value}')
+
+
+@monitoring.command('db_connection_counts')
+@with_appcontext
+def db_connection_counts():
+    """Display DB connection counts."""
+    try:
+        max_conn, used, res_for_super, free = db.session.execute(
+            DB_CONNECTION_COUNTS_QUERY).first()
+    except Exception as error:
+        click.secho(f'ERROR: {error}', fg='red')
+    return click.secho(f'max: {max_conn}, used: {used}, '
+                       f'res_super: {res_for_super}, free: {free}')
+
+
+@monitoring.command('db_connections')
+@with_appcontext
+def db_connections():
+    """Display DB connections."""
+    try:
+        results = db.session.execute(DB_CONNECTION_QUERY).fetchall()
+    except Exception as error:
+        click.secho(f'ERROR: {error}', fg='red')
+    for pid, application_name, client_addr, client_port, backend_start, \
+            xact_start, query_start, wait_event, state, left in results:
+        click.secho(
+            f'application_name: {application_name}\n'
+            f'client_addr: {client_addr}\n'
+            f'client_port: {client_port}\n'
+            f'backend_start: {backend_start}\n'
+            f'xact_start: {xact_start}\n'
+            f'query_start: {query_start}\n'
+            f'wait_event: {wait_event}\n'
+            f'state: {state}\n'
+            f'left: {left}\n'
+        )
